+++ topobank/manager/containers.py
@@ -1,7 +1,7 @@
 """
 Import and export surfaces through archives aka "surface containers".
 """
-
+import datetime
 import json
 import logging
 import math
@@ -12,14 +12,140 @@ import zipfile
 import SurfaceTopography
 import yaml
 from django.conf import settings
+from django.core.files import File
+from django.core.files.storage import default_storage
 from django.utils.text import slugify
 from django.utils.timezone import now
 
 import topobank
 
-from .models import Topography
+from .models import Surface, Topography, topography_datafile_path
 
 _log = logging.getLogger(__name__)
+
+TOPOGRAPHY_DEFAULT_ATTR_VALUES = {
+    'tags': [],
+    'fill_undefined_data_mode': Topography.FILL_UNDEFINED_DATA_MODE_NOFILLING,
+    'detrend_mode': 'center',
+    'is_periodic': False,
+    'height_scale': 1.,
+    'data_source': 0,
+    'measurement_date': datetime.date(1970, 1, 1),
+    'description': '',
+}
+
+SURFACE_DEFAULT_ATTR_VALUES = {
+    'tags': [],
+    'description': '',
+    'is_published': False,
+}
+
+
+def import_topography(topo_dict, topo_file, surface, ignore_missing=False):
+    topo_name = topo_dict['name']
+
+    size_x, *size_rest = topo_dict['size']
+    size_y = None if len(size_rest) == 0 else size_rest[0]
+
+    user = surface.creator
+
+    if ignore_missing:
+        for k, default in TOPOGRAPHY_DEFAULT_ATTR_VALUES.items():
+            topo_dict.setdefault(k, default)
+
+    topo_kwargs = dict(
+        creator=user,
+        name=topo_name,
+        surface=surface,
+        size_x=size_x,
+        size_y=size_y,
+        measurement_date=topo_dict['measurement_date'],
+        description=topo_dict['description'],
+        data_source=topo_dict['data_source'],
+        unit=topo_dict['unit'],
+        tags=topo_dict['tags'],
+        fill_undefined_data_mode=topo_dict['fill_undefined_data_mode'],
+        detrend_mode=topo_dict['detrend_mode'],
+        is_periodic=topo_dict['is_periodic'],
+    )
+
+    try:
+        topo_kwargs.update(dict(
+            instrument_name=topo_dict['instrument']['name'],
+            instrument_type=topo_dict['instrument']['type'],
+            instrument_parameters=topo_dict['instrument']['parameters'],
+        ))
+    except KeyError:
+        # Metadata does not contain instrument information
+        pass
+
+    try:
+        topo_kwargs['height_scale'] = topo_dict['height_scale']
+    except KeyError:
+        # If height_scale is not included, it will probably already
+        # applied because of file contents while loading
+        pass
+
+    topo_kwargs['datafile'] = topo_name
+    topography = Topography(**topo_kwargs)
+
+    # We need to save the topography to get an id...
+    topography.save()
+    # ...which we need for the storage prefix
+    new_topo_file_path = topography_datafile_path(topography, topo_name)
+    actual_topo_file_path = default_storage.save(new_topo_file_path, File(topo_file))
+    topography.datafile = actual_topo_file_path
+    # We need to save again to store the new file name
+    topography.save()
+
+
+def import_container(surface_zip, user, datafile_attribute='datafile', ignore_missing=False, unsafe_yaml_loader=False):
+    """Process surface archive i.e. importing the surfaces.
+
+    Parameters
+    ----------
+    surface_zip : zipfile.Zipfile
+        archive with surfaces
+    user : topobank.users.User
+        User which should be creator of the surface and all data.
+    datafile_attribute : str
+        attribute in file from which the file name for a topography is stored
+    ignore_missing : bool
+        If True, try to find reasonable defaults for missing attributes.
+    unsafe_yaml_loader : bool
+        If True, use an unsafe YAML loader. Use this with care,
+        if there is malicious code in the metadata, this can compromise your system.
+    """
+    surfaces = []
+    with surface_zip.open('meta.yml', mode='r') as meta_file:
+        yaml_loader = yaml.full_load if unsafe_yaml_loader else yaml.safe_load
+        meta = yaml_loader(meta_file)
+
+        for surface_dict in meta['surfaces']:
+            if ignore_missing:
+                for k, default in SURFACE_DEFAULT_ATTR_VALUES.items():
+                    surface_dict.setdefault(k, default)
+
+            import_time = str(now())
+
+            surface_description = surface_dict['description']
+            surface_description += f'\n\nImported from file "{surface_zip.filename}" on {import_time}.'
+            surface = Surface(creator=user,
+                              name=surface_dict['name'],
+                              category=surface_dict['category'],
+                              description=surface_description,
+                              tags=surface_dict['tags'])
+            surface.save()
+
+            surfaces += [surface]
+
+            for topo_idx, topo_dict in enumerate(surface_dict['topographies']):
+                datafile_name = topo_dict[datafile_attribute]['original']  # we just import the original
+                topo_file = surface_zip.open(datafile_name, mode='r')
+
+                import_topography(topo_dict, topo_file, surface, ignore_missing=ignore_missing)
+
+    return surfaces
 
 
 def write_surface_container(file, surfaces):
@@ -48,7 +174,7 @@ def write_surface_container(file, surfaces):
     nb_surfaces = len(surfaces)
     log10_nb_surfaces = int(math.log10(nb_surfaces)) if nb_surfaces > 0 else 0
     for surface_index, surface in enumerate(surfaces):
-        surface_prefix = '' if nb_surfaces == 1 else f'{surface_index}'.zfill(log10_nb_surfaces+1) + '-'
+        surface_prefix = '' if nb_surfaces == 1 else f'{surface_index}'.zfill(log10_nb_surfaces + 1) + '-'
 
         topographies = Topography.objects.filter(surface=surface)
 
@@ -188,7 +314,7 @@ def write_surface_container(file, surfaces):
             - '{}' (legal code), or
             - the included file '{}' (legal code).
             """.format(license_info['title'],
-                       "-"*len(license_info['title']),
+                       "-" * len(license_info['title']),
                        license_info['description_url'],
                        license_info['legal_code_url'],
                        license_file_in_archive))