+++ topobank/manager/models.py
@@ -26,6 +26,7 @@ from django.shortcuts import reverse
 from guardian.models import GroupObjectPermissionBase, UserObjectPermissionBase
 from guardian.shortcuts import assign_perm, get_perms, get_users_with_perms, remove_perm
 from notifications.signals import notify
+from SurfaceTopography.Container.SurfaceContainer import SurfaceContainer
 from SurfaceTopography.Exceptions import UndefinedDataError
 from SurfaceTopography.Support.UnitConversion import get_unit_conversion_factor
 
@@ -76,16 +77,6 @@ class DZIGenerationException(ThumbnailGenerationException):
     pass
 
 
-class TagModel(tm.TagTreeModel):
-    """This is the common tag model for surfaces and topographies.
-    """
-
-    class TagMeta:
-        force_lowercase = True
-        # not needed yet
-        # autocomplete_view = 'manager:autocomplete-tags'
-
-
 class SubjectMixin:
     """Extra methods common to all instances which can be subject to an analysis.
     """
@@ -116,9 +107,14 @@ class SubjectMixin:
         raise NotImplementedError()
 
     def related_surfaces(self):
-        """Returns sequence of related surfaces.
+        """Returns a list of related surfaces. This can be either the parent
+        surface (for a topography), the child surfaces (for a tag), or the
+        surface itself (for a surface).
 
-        :return: True or False
+        Returns
+        -------
+        surfaces : list of Surface
+            The surfaces that are related to this object.
         """
         raise NotImplementedError()
 
@@ -127,13 +123,47 @@ class SubjectMixin:
 
         Returns
         -------
-        A queryset of users.
+        users : Queryset of users
+            All users that have some form of permission (view, full, etc.) on
+            this subject.
         """
         return User.objects.intersection(*tuple(get_users_with_perms(s) for s in self.related_surfaces()))
 
 
-class Surface(models.Model, SubjectMixin):
-    """Physical Surface.
+class Tag(tm.TagTreeModel,
+          SubjectMixin):
+    """This is the common tag model for surfaces and topographies.
+    """
+
+    class TagMeta:
+        force_lowercase = True
+        # not needed yet
+        # autocomplete_view = 'manager:autocomplete-tags'
+
+    def is_shared(self, with_user, allow_change=False):
+        return True  # Tags are generally shared, but the surfaces may not
+
+    def related_surfaces(self):
+        return list(Surface.objects.filter(tags=self.id))
+
+
+class TopobankLazySurfaceContainer(SurfaceContainer):
+    """Wraps a `Surface` with lazy loading of topography data"""
+    def __init__(self, surface):
+        self._surface = surface
+        self._topographies = self._surface.topography_set.all()
+
+    def __len__(self):
+        return len(self._topographies)
+
+    def __getitem__(self, item):
+        return self._topographies[item].read()
+
+
+class Surface(models.Model,
+              SubjectMixin):
+    """
+    A physical surface of a specimen.
 
     There can be many topographies (measurements) for one surface.
     """
@@ -149,7 +179,7 @@ class Surface(models.Model, SubjectMixin):
     creator = models.ForeignKey(User, on_delete=models.CASCADE)
     description = models.TextField(blank=True)
     category = models.CharField(max_length=3, choices=CATEGORY_CHOICES, null=True, blank=False)
-    tags = tm.TagField(to=TagModel)
+    tags = tm.TagField(to=Tag)
     creation_datetime = models.DateTimeField(auto_now_add=True, null=True)
     modification_datetime = models.DateTimeField(auto_now=True, null=True)
 
@@ -350,6 +380,14 @@ class Surface(models.Model, SubjectMixin):
         """
         return hasattr(self, 'publication')  # checks whether the related object surface.publication exists
 
+    def lazy_read(self):
+        """
+        Returns a `SurfaceTopography.Container.SurfaceContainer`
+        representation of this dataset. Reading of actual data is deferred
+        to the point where it is actually needed.
+        """
+        return TopobankLazySurfaceContainer(self)
+
 
 class SurfaceUserObjectPermission(UserObjectPermissionBase):
     content_object = models.ForeignKey(Surface, on_delete=models.CASCADE)
@@ -357,39 +395,6 @@ class SurfaceUserObjectPermission(UserObjectPermissionBase):
 
 class SurfaceGroupObjectPermission(GroupObjectPermissionBase):
     content_object = models.ForeignKey(Surface, on_delete=models.CASCADE)
-
-
-class SurfaceCollection(models.Model, SubjectMixin):
-    """A collection of surfaces."""
-    MAX_LENGTH_NAME = 160
-
-    name = models.CharField(max_length=MAX_LENGTH_NAME)
-    surfaces = models.ManyToManyField(Surface)
-
-    # We have a manytomany field, because a surface could be part of multiple collections.
-
-    @property
-    def label(self):
-        return self.name
-
-    def related_surfaces(self):
-        return list(self.surfaces.all())
-
-    def is_shared(self, with_user, allow_change=False):
-        """Returns True, if this subject is shared with a given user.
-
-        Always returns True if user is the creator of all related surfaces.
-
-        Parameters
-        ----------
-        with_user: User
-            User to test
-
-        Returns
-        -------
-        True or False
-        """
-        return all(s.is_shared(with_user, allow_change=allow_change) for s in self.related_surfaces())
 
 
 def topography_datafile_path(instance, filename):
@@ -405,7 +410,9 @@ def topography_thumbnail_path(instance, filename):
 
 
 class Topography(TaskStateModel, SubjectMixin):
-    """Topography measurement of a surface."""
+    """
+    A single topography measurement of a surface of a specimen.
+    """
 
     # TODO After upgrade to Django 2.2, use constraints: https://docs.djangoproject.com/en/2.2/ref/models/constraints/
     class Meta:
@@ -466,7 +473,7 @@ class Topography(TaskStateModel, SubjectMixin):
     creator = models.ForeignKey(User, null=True, on_delete=models.SET_NULL)
     measurement_date = models.DateField(null=True, blank=True)
     description = models.TextField(blank=True)
-    tags = tm.TagField(to=TagModel)
+    tags = tm.TagField(to=Tag)
     creation_datetime = models.DateTimeField(auto_now_add=True, null=True)
     modification_datetime = models.DateTimeField(auto_now=True, null=True)
 
@@ -798,7 +805,7 @@ class Topography(TaskStateModel, SubjectMixin):
         if not _IN_CELERY_WORKER_PROCESS and self.size_y is not None:
             _log.warning('You are requesting to load a (2D) topography and you are not within in a Celery worker '
                          'process. This operation is potentially slow and may require a lot of memory - do not use '
-                         '`Topography.topography` within the main Django server!')
+                         '`Topography.read` within the main Django server!')
 
         reader_kwargs = dict(channel_index=self.data_source,
                              periodic=self.is_periodic)
@@ -834,7 +841,7 @@ class Topography(TaskStateModel, SubjectMixin):
             topo = topo.interpolate_undefined_data(self.fill_undefined_data_mode)
         return topo.detrend(detrend_mode=self.detrend_mode)
 
-    def topography(self, allow_squeezed=True, return_reader=False):
+    def read(self, allow_squeezed=True, return_reader=False):
         """Return a SurfaceTopography.Topography/UniformLineScan/NonuniformLineScan instance.
 
         This instance is guaranteed to
@@ -875,7 +882,7 @@ class Topography(TaskStateModel, SubjectMixin):
                 _log.warning(
                     'You are requesting to load a (2D) topography and you are not within in a Celery worker '
                     'process. This operation is potentially slow and may require a lot of memory - do not use '
-                    '`Topography.topography` within the main Django server!')
+                    '`Topography.read` within the main Django server!')
 
             # Okay, we can use the squeezed datafile, it's already there.
             toporeader = get_topography_reader(self.squeezed_datafile, format=SQUEEZED_DATAFILE_FORMAT)
@@ -894,6 +901,8 @@ class Topography(TaskStateModel, SubjectMixin):
             return topo, toporeader
         else:
             return topo
+
+    topography = read  # Renaming this, mark `topography` as deprecated before v2
 
     def to_dict(self):
         """Create dictionary for export of metadata to json or yaml"""
@@ -982,7 +991,7 @@ class Topography(TaskStateModel, SubjectMixin):
             Thumbnail image.
         """
         if st_topo is None:
-            st_topo = self.topography()  # SurfaceTopography instance (=st)
+            st_topo = self.read()  # SurfaceTopography instance (=st)
         image_file = io.BytesIO()
         if st_topo.dim == 1:
             dpi = 100
@@ -1024,7 +1033,7 @@ class Topography(TaskStateModel, SubjectMixin):
         None
         """
         if st_topo is None:
-            st_topo = self.topography()
+            st_topo = self.read()
 
         image_file = self.get_thumbnail(st_topo=st_topo)
 
@@ -1048,7 +1057,7 @@ class Topography(TaskStateModel, SubjectMixin):
         None
         """
         if st_topo is None:
-            st_topo = self.topography()
+            st_topo = self.read()
         if self.size_y is not None:
             # This is a topography (map), we need to create a Deep Zoom Image
             make_dzi(st_topo, self._dzi_storage_prefix())
@@ -1112,7 +1121,7 @@ class Topography(TaskStateModel, SubjectMixin):
 
     def renew_squeezed_datafile(self, st_topo=None):
         if st_topo is None:
-            st_topo = self.topography()
+            st_topo = self.read()
         with tempfile.NamedTemporaryFile() as tmp:
             # Write and upload NetCDF file
             st_topo.to_netcdf(tmp.name)
@@ -1132,7 +1141,7 @@ class Topography(TaskStateModel, SubjectMixin):
         Cache bandwidth for bandwidth plot in database. Data is stored in units of meter.
         """
         if st_topo is None:
-            st_topo = self.topography()
+            st_topo = self.read()
         if st_topo.unit is not None:
             bandwidth_lower, bandwidth_upper = st_topo.bandwidth()
             fac = get_unit_conversion_factor(st_topo.unit, 'm')