+++ topobank/manager/urls.py
@@ -6,8 +6,9 @@ from rest_framework.routers import DefaultRouter
 from . import views
 
 router = DefaultRouter()
-router.register(r'api/surface', views.SurfaceViewSet, basename='surface-api')
+router.register(r'api/tag', views.TagViewSet, basename='tag-api')
 router.register(r'api/topography', views.TopographyViewSet, basename='topography-api')
+router.register(r'api/surface', views.SurfaceViewSet, basename='surface-api')
 
 urlpatterns = router.urls
 
@@ -42,6 +43,11 @@ urlpatterns += [
         'api/surface/<pk>/set-permissions/',
         view=login_required(views.set_permissions),
         name='set-permissions'
+    ),
+    path(
+        'api/import-surface/',
+        view=login_required(views.import_surface),
+        name='import-surface'
     ),
     path(
         'api/versions/',