+++ topobank/manager/views.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.core.exceptions import PermissionDenied
 from django.core.files.storage import default_storage
 from django.db.models import Prefetch, Q
-from django.http import Http404, HttpResponse, HttpResponseForbidden
+from django.http import Http404, HttpResponse, HttpResponseBadRequest, HttpResponseForbidden
 from django.shortcuts import redirect
 from django.utils.text import slugify
 from guardian.shortcuts import assign_perm, get_users_with_perms, remove_perm
@@ -22,12 +22,20 @@ from ..taskapp.utils import run_task
 from ..usage_stats.utils import increase_statistics_by_date_and_object
 from ..users.models import User
 from .containers import write_surface_container
-from .models import Surface, Topography, topography_datafile_path
+from .models import Surface, Tag, Topography, topography_datafile_path
 from .permissions import ObjectPermissions, ParentObjectPermissions
-from .serializers import SurfaceSerializer, TopographySerializer
+from .serializers import SurfaceSerializer, TagSerializer, TopographySerializer
+from .tasks import import_container_from_url
 from .utils import api_to_guardian, get_upload_instructions
 
 _log = logging.getLogger(__name__)
+
+
+class TagViewSet(mixins.RetrieveModelMixin,
+                 viewsets.GenericViewSet):
+    queryset = Tag.objects.all()
+    serializer_class = TagSerializer
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
 
 class SurfaceViewSet(mixins.CreateModelMixin,
@@ -313,6 +321,19 @@ def upload_topography(request, pk=None):
 
     # Return 204 No Content
     return Response({}, status=204)
+
+
+@api_view(['POST'])
+def import_surface(request):
+    url = request.data.get('url')
+
+    if not url:
+        return HttpResponseBadRequest()
+
+    user = request.user
+    import_container_from_url.delay(user.id, url)
+
+    return Response({}, status=200)
 
 
 @api_view(['GET'])