+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for *TopoBank*
 
+## 1.7.0 (not yet released)
+
+- ENH: Tags as analysis subjects (as a replacement for surface collections)
+- ENH: Import of ZIP containers from URL
+
 ## 1.6.1 (2024-01-30)
 
 - MAINT: Updated to django-allauth >= 0.56.0